+++ packages/stack-ts/src/index.ts
@@ -2,3 +2,4 @@ export * from "./stack-config.js";
 export * from "./parsers.js";
 export * from "./discover.js";
 export * from "./run.js";
+export * from "./templates.js";